Initial commit
Python script to generate a macro that compiles an embedded c# program and runs an embedded c# program via installutil. The C# program calls an embedded version of the invoke-shellcode script. The invoke-shellcode script calls shellcode that is generated by msfvenom via the python script.

Based on the work of many people (subTee, mattifestation, and others)